ledger: move migrations into a storage package (#4846)
