+++ util/db/versioning.go
@@ -26,9 +26,9 @@ import (
 
 // GetUserVersion returns the user version field stored in the sqlite database
 // if the database was never initiliazed with a version, it would return 0 as the version.
-func GetUserVersion(ctx context.Context, tx *sql.Tx) (userVersion int32, err error) {
+func GetUserVersion(ctx context.Context, q Queryable) (userVersion int32, err error) {
 
-	err = tx.QueryRowContext(ctx, "PRAGMA user_version").Scan(&userVersion)
+	err = q.QueryRowContext(ctx, "PRAGMA user_version").Scan(&userVersion)
 	// it's not really supposed to happen with a user_version, since the above would always succeed, but
 	// we want to have it so that we can align with the SQL statements "correct handling practices".
 	if err == sql.ErrNoRows {
@@ -39,15 +39,15 @@ func GetUserVersion(ctx context.Context, tx *sql.Tx) (userVersion int32, err err
 }
 
 // SetUserVersion sets the userVersion as the new user version, and return the old version.
-func SetUserVersion(ctx context.Context, tx *sql.Tx, userVersion int32) (previousUserVersion int32, err error) {
-	previousUserVersion, err = GetUserVersion(ctx, tx)
+func SetUserVersion(ctx context.Context, e Executable, userVersion int32) (previousUserVersion int32, err error) {
+	previousUserVersion, err = GetUserVersion(ctx, e)
 	if err != nil {
 		return
 	}
 	if previousUserVersion == userVersion {
 		return
 	}
-	_, err = tx.ExecContext(ctx, fmt.Sprintf("PRAGMA user_version = %d", userVersion))
+	_, err = e.ExecContext(ctx, fmt.Sprintf("PRAGMA user_version = %d", userVersion))
 	if err != nil {
 		// if we're aborting due to an error, clear the previousUserVersion so that
 		// on all error cases we'll be returning zero.
